Productivity_App.py - a graphical interface that keeps track of your activities and time spent in doing the task. This was created for personal use and for familiarization of the tkinter module, as well as threading. Honestly didn't think that this would take up much more time than expected - I was unable to foresee the complexity of the timers and switching between multiple frames. But good practice is never easy!
Initial Release
- GUI made through Tkinter. Frames created for mainframe,
pacedframe, timedframe, pomodoroframe and pacedlobbyframe.

- GUI styling made flexible, so that styling can be changed
depending on the user.

- Added elements for mainframe, pacedframe and pacedlobby.

- Paced Productivity - sets a timer for you. No limits and
no minimum time for productivity, just at your own pace! You
can set the description of the activity, and then proceed to
start doing it. If you need to take breaks, a pause button is
available as well. Note that paced_pause() can only be initiated
once per second, so the timer has to wait for the next second
after pausing for it to be actually paused. This can be made
a bit more accurate by adding another counter for milliseconds.

- Clicking the End Session button will initiate a file write that
writes a log in prod.data regarding your recent activity. Note that
closing the application manually using the close button during
productivity sessions will not log the current session. You have to
end session for it to be recorded

- Features to be added are for timed and pomodoro-based productivity.