Productivity App - Timer-Based
- Added the timer-based productivity functionality for the application.
User will input the description of the productity, along with the
proposed time duration. The user can opt to finish the session early
by clicking the end now button, or the user can wait for the program to
expire naturally.

- Couple of things to add will be a sound prompt when the timer finishes
so that the user will be alerted upon time completion.