productivity_app.py - Pomodoro
- added pomodoro technique for productivity on the application. Accepts
inputs of respective time and break times, as well as the type of
activity that is needed to be done. Can be stopped at any time you feel
like you should stop. Will infinitely roll over breaks and activities
until you end the session.

- pause used to be iffy. Since it has to wait for 1 second before it
receives the signal to interrupt, it can possibly ignore pauses and
"speed up" time by continuously clicking on the pause and start buttons
repeatedly. To fix this, we disable all the buttons on the instance of
the pause button click, and will enable the needed buttons only after
the pause has been accepted and done.

- logs for each productivity are saved on a file named prod.data. This
file contains starting datetime, ending datetime, type of productivity,
total time of productivity and description of the productivity.

- next features to create are:
1. prod.data - for entries with different starting date and end date,
(ex. started on sept 12 11:30pm and ended on sept 13 12:30am), we should
find a way to cut the entry into 2. Problem would be the division of the
time duration because of possible pauses.

2. data statistics label - label on the main should include the total time
spent doing productivity for the day. something like: get_currentdate(),
find it on prod.data, add the HMS together.

3. additional statistics - commit another frame for statistics, which can
display weekly/monthly productivity times, greatest productivity day and
least productivity day.