productivity_app.py - data_verify, statistics
- Problem: prod.data - for entries with different starting date
and end date,(ex. started on sept 12 11:30pm and ended on sept 13
12:30am), we should find a way to cut the entry into 2. Problem would
be the division of the time duration because of possible pauses.

- Solution - add a verification checker that checks if the starting
day and ending day of the productivity is the same. If it is not, we
split the data entry into 2 - 1 for each day. Productivity rate occupies
the first day first, then remaining productivity rate will be added on
the next day. We do this so that statistics feature to be added in the
future won't repeat data. Additional problem would be if the entry has
a discrepancy of more than 1 day. (started sept 12, ended sept 15). For
this, we request manual changing of data in prod.data, with error as
seen on the app.

- Feature for daily statistics was added. Main menu now shows you the
current productivity rate you have done for the day.

- Next Features:
1. additional statistics - commit another frame for statistics, which can
display weekly/monthly productivity times, greatest productivity day and
least productivity day.