Productivity_app.py - Beta Testing
- App now contains all basic functionalities needed. Features such
as sounds for timers and pomodoro break/active times and indicators
are now available. Sound is contained in the sounds folder, which
are samples mostly downloaded from freesound, while others were
generated using the gtts module.

- Concern regarding averages was considered. Now we take into account
2 averages, 1 would be averaging on all days of the week/month and
another would be to average only on productive days of the week/month.

- Next feature would be to create an executable file from this. I am
excited!